requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
optuna
joblib
anthropic
requests
beautifulsoup4
python-dotenv
shap
playwright
playwright-stealth
